Refactored folder structure to follow architecture design
