+++ Android/indoor_localization/lib/domain/entities/asset.dart
@@ -1,9 +1,9 @@
-import 'package:indoor_localization/entities/facility.dart';
-import 'package:indoor_localization/entities/point.dart';
-import 'package:indoor_localization/entities/zone.dart';
-import 'package:indoor_localization/services/asset_service.dart';
-import 'package:indoor_localization/services/facility_service.dart';
-import 'package:indoor_localization/services/zone_service.dart';
+import 'package:indoor_localization/domain/entities/facility.dart';
+import 'package:indoor_localization/domain/entities/point.dart';
+import 'package:indoor_localization/domain/entities/zone.dart';
+import 'package:indoor_localization/domain/services/asset_service.dart';
+import 'package:indoor_localization/domain/services/facility_service.dart';
+import 'package:indoor_localization/domain/services/zone_service.dart';
 
 class Asset {
   int id;
