+++ Android/indoor_localization/lib/domain/entities/asset_position_history_log_filter.dart
