+++ Android/indoor_localization/lib/domain/entities/asset_zone_history_log_filter.dart
