+++ Android/indoor_localization/lib/domain/entities/point.dart
