+++ Android/indoor_localization/lib/domain/mock-repositories/mock_asset_position_history_log_repository.dart
@@ -1,6 +1,6 @@
-import 'package:indoor_localization/entities/asset_position_history_log.dart';
-import 'package:indoor_localization/entities/asset_position_history_log_filter.dart';
-import 'package:indoor_localization/repository-interfaces/asset_position_history_log_repository.dart';
+import 'package:indoor_localization/domain/entities/asset_position_history_log.dart';
+import 'package:indoor_localization/domain/entities/asset_position_history_log_filter.dart';
+import 'package:indoor_localization/domain/repository-interfaces/asset_position_history_log_repository.dart';
 
 class MockAssetPositionHistoryLogRepository implements IAssetPositionHistoryLogRepository {
   static List<AssetPositionHistoryLog> logs = [];
