+++ Android/indoor_localization/lib/domain/repository-interfaces/asset_repository.dart
@@ -1,0 +1,5 @@
+import 'package:indoor_localization/domain/entities/asset.dart';
+import 'package:indoor_localization/domain/repository-interfaces/repository.dart';
+
+abstract class IAssetRepository extends IRepository<Asset> {
+}
