+++ Android/indoor_localization/lib/domain/repository-interfaces/asset_zone_history_log_repository.dart
@@ -1,0 +1,7 @@
+import 'package:indoor_localization/domain/entities/asset_zone_history_log.dart';
+import 'package:indoor_localization/domain/entities/asset_zone_history_log_filter.dart';
+import 'package:indoor_localization/domain/repository-interfaces/repository.dart';
+
+abstract class IAssetZoneHistoryLogRepository extends IRepository<AssetZoneHistoryLog> {
+  List<AssetZoneHistoryLog> getLogs(AssetZoneHistoryLogFilter assetZoneHistoryLogFilter);
+}
