+++ Android/indoor_localization/lib/domain/repository-interfaces/facility_repository.dart
@@ -1,0 +1,5 @@
+import 'package:indoor_localization/domain/entities/facility.dart';
+import 'package:indoor_localization/domain/repository-interfaces/repository.dart';
+
+abstract class IFacilityRepository extends IRepository<Facility> {
+}
