+++ Android/indoor_localization/lib/domain/repository-interfaces/repository.dart
