+++ Android/indoor_localization/lib/domain/repository-interfaces/zone_repository.dart
@@ -1,0 +1,5 @@
+import 'package:indoor_localization/domain/entities/zone.dart';
+import 'package:indoor_localization/domain/repository-interfaces/repository.dart';
+
+abstract class IZoneRepository extends IRepository<Zone> {
+}
