+++ Android/indoor_localization/lib/domain/services/asset_service.dart
@@ -1,8 +1,8 @@
-import 'package:indoor_localization/entities/asset.dart';
-import 'package:indoor_localization/entities/facility.dart';
-import 'package:indoor_localization/entities/point.dart';
-import 'package:indoor_localization/entities/zone.dart';
-import 'package:indoor_localization/services/entity_service.dart';
+import 'package:indoor_localization/domain/entities/asset.dart';
+import 'package:indoor_localization/domain/entities/facility.dart';
+import 'package:indoor_localization/domain/entities/point.dart';
+import 'package:indoor_localization/domain/entities/zone.dart';
+import 'package:indoor_localization/domain/services/entity_service.dart';
 
 class AssetService extends EntityService {
   static Asset? get(int id) {
