--- Android/indoor_localization/lib/repository-interfaces/asset_repository.dart
@@ -1,5 +1,0 @@
-import 'package:indoor_localization/entities/asset.dart';
-import 'package:indoor_localization/repository-interfaces/repository.dart';
-
-abstract class IAssetRepository extends IRepository<Asset> {
-}
