--- Android/indoor_localization/lib/repository-interfaces/asset_zone_history_log_repository.dart
@@ -1,7 +1,0 @@
-import 'package:indoor_localization/entities/asset_zone_history_log.dart';
-import 'package:indoor_localization/entities/asset_zone_history_log_filter.dart';
-import 'package:indoor_localization/repository-interfaces/repository.dart';
-
-abstract class IAssetZoneHistoryLogRepository extends IRepository<AssetZoneHistoryLog> {
-  List<AssetZoneHistoryLog> getLogs(AssetZoneHistoryLogFilter assetZoneHistoryLogFilter);
-}
