--- Android/indoor_localization/lib/repository-interfaces/facility_repository.dart
@@ -1,5 +1,0 @@
-import 'package:indoor_localization/entities/facility.dart';
-import 'package:indoor_localization/repository-interfaces/repository.dart';
-
-abstract class IFacilityRepository extends IRepository<Facility> {
-}
