--- Android/indoor_localization/lib/repository-interfaces/zone_repository.dart
@@ -1,5 +1,0 @@
-import 'package:indoor_localization/entities/zone.dart';
-import 'package:indoor_localization/repository-interfaces/repository.dart';
-
-abstract class IZoneRepository extends IRepository<Zone> {
-}
